+++ src/main/java/com/apiscrum/APIScrum/Service/Impl/UserStoryServiceImpl.java
@@ -1,6 +1,12 @@
 package com.apiscrum.APIScrum.Service.Impl;
 
+import com.apiscrum.APIScrum.DTO.UserStoryDto;
+import com.apiscrum.APIScrum.Entity.Epic;
+import com.apiscrum.APIScrum.Entity.ProductBackLog;
 import com.apiscrum.APIScrum.Entity.UserStory;
+import com.apiscrum.APIScrum.Mapper.UserStoryMapper;
+import com.apiscrum.APIScrum.Repository.EpicRepository;
+import com.apiscrum.APIScrum.Repository.ProductBackLogRepository;
 import com.apiscrum.APIScrum.Repository.UserStoryRepository;
 import com.apiscrum.APIScrum.Service.UserStoryService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,20 +16,28 @@ import org.springframework.stereotype.Service;
 public class UserStoryServiceImpl implements UserStoryService {
     @Autowired
     UserStoryRepository userStoryRepository;
+    @Autowired
+    ProductBackLogRepository productBackLogRepository;
+    @Autowired
+    EpicRepository epicRepository;
 
-    @Override
-    public void createUserStory(UserStory userStory) {
-        userStoryRepository.save(userStory);
-    }
+
+
+
 
     @Override
     public void deleteUserStory(Long id) {
-        userStoryRepository.deleteById(id);
+        if(userStoryRepository.existsById(id))
+            userStoryRepository.deleteById(id);
+        else
+            throw new RuntimeException("User Story not found with id: " + id);
     }
 
     @Override
-    public UserStory getUserStory(Long id) {
-        return userStoryRepository.findById(id).get();
+    public UserStoryDto getUserStory(Long id) {
+        UserStory us = userStoryRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User Story not found with id: " + id));
+        return UserStoryMapper.mapToUserStoryDTO(us);
     }
 
 }
